routes and links are up and running
