--- components/Header_Components/AddRestaurant.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-
-
-function AddRestaurant () {
-return( 
-   <Link href="/RestForm" ><button type="submit">Add Restaurant</button></Link>
-)
-}
-
-export default AddRestaurant
