--- components/Restaurant_Form_Components/FormSubmit.js
@@ -1,15 +1,0 @@
-import React from 'react'
-
-
-
-function FormSubmit (props) {
-//if zip is NaN do not send
-function handleClick() {
-console.log(props)
-}
-
-return(
-    <Link href="/"><button onClick={handleClick} type="submit">Add Restraunt</button></Link>
-)}
-
-export default FormSubmit
