+++ src/components/Header.js
@@ -3,7 +3,7 @@ import SearchKey from './Header_Components/SearchKey'
 import SearchLocation from './Header_Components/SearchLocation'
 import SearchButton from './Header_Components/SearchButton'
 import AddRestaurant from './Header_Components/AddRestaurant'
-
+import { Link } from 'react-router-dom'
 
 
 
@@ -11,9 +11,8 @@ import AddRestaurant from './Header_Components/AddRestaurant'
 function Header (props) {
 
         return (
-            <>
             <nav id='home'>
-                <Link href="/"><img src= 'https://s3-media0.fl.yelpcdn.com/assets/public/default.yelp_design_web.yji-b085a608c15f110ce750ccef3e1e1db0.png' alt='yelp'/></Link>
+                <Link to='/'><img src= 'https://s3-media0.fl.yelpcdn.com/assets/public/default.yelp_design_web.yji-b085a608c15f110ce750ccef3e1e1db0.png' alt='yelp'/></Link>
                 <SearchButton
                 />
                 <SearchLocation
@@ -22,16 +21,9 @@ function Header (props) {
                 <SearchKey
                     onChange={props.onChange}
                 />
-                <AddRestraunt
+                <AddRestaurant
                 />
             </nav>
-            <style jsx>{`
-                #search{ color: red; }
-            
-            `}
-
-            </style>
-            </>
         )
     }
