+++ src/components/Header_Components/AddRestaurant.js
@@ -1,0 +1,11 @@
+import React from 'react'
+import { Link } from 'react-router-dom'
+
+
+function AddRestaurant () {
+return( 
+   <Link to='/Add'>  <button type="submit">Add Restaurant</button></Link>
+)
+}
+
+export default AddRestaurant
