+++ src/components/Header_Components/SearchButton.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import { Link } from 'react-router-dom'
+
+
+function SearchButton(props){
+return(
+<div id='search'>
+    <Link to='/Search/:name'> <button onClick={props.onChange} type="search" >Search</button> </Link>
+ </div>
+)}
+
+
+export default SearchButton
