+++ src/components/Restaurant_Form_Components/FormName.js
@@ -1,0 +1,20 @@
+import React from 'react'
+
+
+
+function FormName (props) {
+
+    //  handleChange =(e)=>{
+    //     let name = e.target.value
+    //     props.newName(name)
+    //      onChange = { this.handleChange }
+    // }
+
+return(
+    <div>
+        <h2>Form Name</h2>
+        <input type="text" placeholder="Form Name"  name="search" />
+    </div>
+)}
+
+export default FormName
