+++ src/components/Restaurant_Form_Components/FormSubmit.js
@@ -1,0 +1,17 @@
+import React from 'react'
+import { Link } from 'react-router-dom'
+
+
+function FormSubmit (props) {
+//if zip is NaN do not send
+// function handleClick() {
+// console.log(props)
+// }
+//     onClick = { this.handleClick }
+
+
+return(
+    <Link href="/"><button type="submit">Add Restraunt</button></Link>
+)}
+
+export default FormSubmit
