+++ src/components/Review_form.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
-import Star_Review from './Review_Form/Star_Review'
+import Star_Review from './Review_Form/Star_Reviews'
 import Review_Field from './Review_Form/Review_Field'
-import Link from 'next/link'
+
 
 function Review_Form () {
 
@@ -27,7 +27,7 @@ return(
         <Review_Field 
             newText={newText}
 />
-        <Link href="/"><button type="submit">Submit</button></Link>
+        <button type="submit">Submit</button>
     </nav>
 )}
