+++ src/components/SearchRes.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import Results from './Search/Results'
-import Link from 'next/link'
+
 
 function SearchRes (props) {
 return(
